Move Lin into the BW folder as well
